requirements.txt
###########################################################################################

#BACKEND:
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.6
filelock==3.16.1
flatbuffers==25.1.21
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==10.2.0
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==70.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2


########################################################################################
# Frontend
Γö£ΓöÇΓöÇ @types/react-router-dom@5.3.3
Γö£ΓöÇΓöÇ autoprefixer@10.4.20
Γö£ΓöÇΓöÇ axios@1.7.9
Γö£ΓöÇΓöÇ chart.js@4.4.7
Γö£ΓöÇΓöÇ cra-template-typescript@1.2.0
Γö£ΓöÇΓöÇ postcss@8.5.1
Γö£ΓöÇΓöÇ react-chartjs-2@5.3.0
Γö£ΓöÇΓöÇ react-dom@19.0.0
Γö£ΓöÇΓöÇ react-router-dom@7.1.3
Γö£ΓöÇΓöÇ react-scripts@5.0.1
Γö£ΓöÇΓöÇ react@19.0.0
ΓööΓöÇΓöÇ tailwindcss@3.4.17